refactor: timer 기능 리팩토링
