+++ client/src/pages/study_room/study_room.jsx
@@ -19,14 +19,6 @@ function StudyRoom() {
   const [isSidebar, setisSidebar] = useState(false);
   const displayType = isSidebar === false ? styles.hide : styles.show;
 
-  const [startTime, setStartTime] = useState(null);
-  const [endTime, setEndTime] = useState(null);
-  const [isRecorded, setIsRecorded] = useState(false);
-
-  const [hour, setHour] = useState(0);
-  const [minute, setMinute] = useState(0);
-  const [second, setSecond] = useState(0);
-
   const leaveRoom = () => {
     session.disconnect();
     setSubscribers([]);
@@ -90,41 +82,6 @@ function StudyRoom() {
     }
   }, [session]);
 
-  useEffect(() => {
-    let timer;
-    if (isRecorded) {
-      timer = setInterval(() => {
-        if (minute > 59) {
-          setHour((prev) => prev + 1);
-          setMinute(0);
-        }
-        if (second === 59) {
-          setMinute((prev) => prev + 1);
-          setSecond(0);
-        }
-        if (second < 59) {
-          setSecond((prev) => prev + 1);
-        }
-      }, 1000);
-    } else {
-      clearInterval(timer);
-    }
-    return () => clearInterval(timer);
-  });
-
-  const getStartTime = () => {
-    const time = new Date();
-    setStartTime(time);
-    setIsRecorded((prev) => !prev);
-  };
-
-  const getEndTime = () => {
-    const time = new Date();
-    setEndTime(time);
-    setIsRecorded((prev) => !prev);
-    /* server에 selectedPlan, startTime, endTime, studyRooomId post */
-  };
-
   const clickPlanBtn = () => {
     setClickedPlanBtn((prev) => !prev);
     setisSidebar((prev) => !prev);
@@ -144,13 +101,8 @@ function StudyRoom() {
     <div className={styles.room}>
       <div className={styles.video_container}>
         <ul className={`${styles.videos} ${displayType}`}>
-          {publisher && (
-            <UserVideo count={count} streamManager={publisher} hour={hour} minute={minute} second={second} />
-          )}
-          {subscribers &&
-            subscribers.map((subscriber) => (
-              <UserVideo count={count} streamManager={subscriber} hour={hour} minute={minute} second={second} />
-            ))}
+          {publisher && <UserVideo count={count} streamManager={publisher} />}
+          {subscribers && subscribers.map((subscriber) => <UserVideo count={count} streamManager={subscriber} />)}
         </ul>
         <div className={`${styles.plan_bar} ${getStyles(clickedPlanBtn)}`}>
           <PlanSidebar />
@@ -161,9 +113,6 @@ function StudyRoom() {
           toggleVideo={toggleVideo}
           toggleAudio={toggleAudio}
           leaveRoom={leaveRoom}
-          isRecorded={isRecorded}
-          getStartTime={getStartTime}
-          getEndTime={getEndTime}
           onClickplanBtn={clickPlanBtn}
         />
       </div>
